App/requirements.txt
# Framework Principal
fastapi==0.115.5
uvicorn==0.32.0

# Base de Datos y ORM
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.2

# Autenticación y Seguridad
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.17
bcrypt==4.2.0
PyJWT==2.10.1

# Manejo de Archivos y Cloud
python-dotenv==1.0.1
cloudinary==1.41.0
Pillow==10.3.0

# Utilidades HTTP y API
requests==2.31.0
httpx==0.27.2
python-dateutil==2.8.2

# Validación y Serialización
pydantic==2.9.2
email-validator==2.2.0

# CORS y Middleware
starlette==0.41.2

# Utilidades Generales
python-multipart==0.0.17
typing_extensions==4.12.2